+++ docs-markdown/src/controllers/bold-controller.ts
@@ -4,13 +4,14 @@ import { Range, Selection, TextEditorEdit, window } from 'vscode';
 import {
 	insertContentToEditor,
 	isMarkdownFileCheck,
+	isValidEditor,
 	noActiveEditorMessage,
 	postWarning,
 	showStatusMessage
 } from '../helper/common';
 import { insertUnselectedText } from '../helper/format-logic-manager';
 import { isBold, isBoldAndItalic } from '../helper/format-styles';
-import { reporter, sendTelemetryData } from '../helper/telemetry';
+import { sendTelemetryData } from '../helper/telemetry';
 
 const telemetryCommand: string = 'formatBold';
 
@@ -20,15 +21,18 @@ export function boldFormattingCommand() {
 }
 
 /**
- * Replaces current selection with MD bold formated selection
+ * Replaces current selection with MD bold formatted selection
  */
 export function formatBold() {
-	reporter.sendTelemetryEvent(`${telemetryCommand}`);
 	const editor = window.activeTextEditor;
 	if (!editor) {
 		noActiveEditorMessage();
 		return;
 	} else {
+		if (!isValidEditor(editor, true, 'format bold')) {
+			return;
+		}
+
 		if (!isMarkdownFileCheck(editor, false)) {
 			return;
 		}
@@ -70,7 +74,7 @@ export function formatBold() {
 			insertContentToEditor(editor, formattedText, true);
 		}
 
-		// if mulitple cursors were used to make selections
+		// if multiple cursors were used to make selections
 		if (selections.length > 1) {
 			editor
 				.edit((edit: TextEditorEdit): void => {
@@ -107,7 +111,7 @@ export function bold(content: string, range?: Range) {
 		return selectedText.substring(2, selectedText.length - 2);
 	}
 
-	// Set sytax for bold formatting and replace original string with formatted string
+	// Set syntax for bold formatting and replace original string with formatted string
 	const styleBold = `**${selectedText}**`;
 	return styleBold;
 }