+++ docs-markdown/changelog.md
@@ -24,6 +24,7 @@
 - Added Unit Test Coverage: code-controller.ts
 - Added Moniker to quick-pick-menu-controller.ts
 - Added Unit Test Coverage: include-controller.ts
+- Added Unit Test Coverage: bold-controller.ts
 
 ## 0.2.63 (April 2nd, 2020)